Fix #631
Always add the list item for the export action to the home screen, but hide it if there are no items in the app's KB collection (this can happen if the call to fetch the KB hasn't returned yet). When the KB collection retrieval finishes, the app looks to see if the Export action is hidden, and if so, displays the action.